Get the async mechanism to work + document some functions.
It was a bit of a slog to figure out mechanisms around async.
While the system itself if quite simple, managing all the memory
to make it work was difficult. Had quite a few segnemtation faults
during the process. It would be either losing the function frame
when caller goes out of scope, or trying to modify variables that
exist in unmodifiable memory. All in all, good learning experience.