Implemented a blinker that utilises the new Scheduler.
Although still a carcass of the targeted result, it provides
a kick-off point for further development. This version still has an error.
The chip raises an access violation exception on compare and swap in
the Scheduler. After reading the manual I've realized that this class
of instruction will do that in DTIM region. I'll have to set up and use
the ITIM memory before I can continue. It seems to be the only region
that should allow these instructions to execute. I'll attempt this
in another commit and hope this works. That instruction is critical
for multi hart/core systems, where there may be a few threads competing.