Fixed the stack overflow troubles of the previous commit.
Increased the stack size. It was quite small before, I did not expect
to use a lot of it this soon. Though troubles did not end. ITIM memory
region is also not suitable for LR/SC instructions, therefore an
access violation is also raised by the MCU, despite it using ITIM.

Started reading the documentation I could find (and online forums).
It's possible that there is a typo in the manual, that's why I interpreted
it wrong. It seems only 'cache-able' regions are ok, ITIM does not count.

Now it actually seems like we do not have any LR/SC operations able
to execute on the MCU. There are no memory regions that are both
write-able and cache-able. A dead-end for this Scheduler feature.
This will have to be made hardware specific, I believe multi-core
MCU's will have no such limitation, because this is a vital capability.